refactor: move image modules into images folder
