+++ src/nrtk_explorer/app/embeddings.py
@@ -6,7 +6,7 @@ from nrtk_explorer.library.dataset import get_dataset
 from nrtk_explorer.app.trame_utils import SetStateAsync
 from nrtk_explorer.app.applet import Applet
 import nrtk_explorer.test_data
-from nrtk_explorer.app.image_ids import (
+from nrtk_explorer.app.images.image_ids import (
     image_id_to_dataset_id,
     dataset_id_to_transformed_image_id,
     dataset_id_to_image_id,
